chore: Adjusting DataType for ControlExtensions
