+++ src/library/Uno.Material/Styles/Controls/v2/TextBox.xaml
@@ -375,7 +375,8 @@
 										<Setter Target="PlaceholderElement.Foreground" Value="{ThemeResource FilledTextBoxPlaceholderForegroundDisabled}" />
 										<Setter Target="Root.Background" Value="{ThemeResource FilledTextBoxBackgroundDisabled}" />
 										<Setter Target="DeleteButton.Foreground" Value="{ThemeResource FilledTextBoxDeleteButtonForegroundDisabled}" />
-										<Setter Target="IconPresenter.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
+										<Setter Target="LeadingIconButton.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
+										<Setter Target="TrailingIconButton.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
 									</VisualState.Setters>
 								</VisualState>
 
@@ -622,7 +623,8 @@
 										<Setter Target="ContentElement.Foreground" Value="{ThemeResource OutlinedTextBoxForegroundDisabled}" />
 										<Setter Target="ContentElement.Opacity" Value="{ThemeResource OutlinedTextBoxForegroundOpacityDisabled}" />
 										<Setter Target="PlaceholderElement.Foreground" Value="{ThemeResource OutlinedTextBoxPlaceholderForegroundDisabled}" />
-										<Setter Target="IconPresenter.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
+										<Setter Target="LeadingIconButton.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
+										<Setter Target="TrailingIconButton.Foreground" Value="{ThemeResource TextBoxLeadingIconForegroundDisabled}" />
 									</VisualState.Setters>
 								</VisualState>
