+++ src/samples/UWP/Uno.Themes.Samples.Shared/Entities/Sample.cs
@@ -1,18 +1,13 @@
 ﻿using System;
-using System.Windows.Input;
 using Uno.Extensions;
 using Uno.Logging;
-using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Data;
-using static Uno.Themes.Samples.ViewModels.ViewModelBase;
 
 namespace Uno.Themes.Samples.Entities
 {
 	[Bindable]
 	public class Sample
 	{
-		public ICommand MyCommand { get; private set; }
-
 		public Sample(SamplePageAttribute attribute, Type viewType)
 		{
 			Category = attribute.Category;
@@ -25,8 +20,6 @@ namespace Uno.Themes.Samples.Entities
 			SortOrder = attribute.SortOrder;
 
 			ViewType = viewType;
-
-			MyCommand = new Command(ExecuteMyCommand);
 		}
 
 		private object CreateData(Type dataType)
@@ -42,23 +35,6 @@ namespace Uno.Themes.Samples.Entities
 				this.Log().Error($"Failed to initialize data for `{ViewType.Name}`:", e);
 				return null;
 			}
-		}
-
-		private async void ShowWarning()
-		{
-			var messageDialog = new ContentDialog
-			{
-				Title = "Command Dialog",
-				Content = "A command was fired",
-
-				CloseButtonText = "Ok"
-			};
-			await messageDialog.ShowAsync();
-		}
-
-		private void ExecuteMyCommand(object parameter)
-		{
-			ShowWarning();
 		}
 
 		public SampleCategory Category { get; }
