+++ src/samples/UWP/Uno.Themes.Samples.Shared/Uno.Themes.Samples.Shared.projitems
@@ -151,6 +151,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)Converters\FromBoolToValueConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Converters\FromStringToValueConverter.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\Data\TestCollections.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Entities\Data\TestCommands.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\Design.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\Icons.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Entities\Sample.cs" />
